feat: implement PID JWT payload claims and SD disclosures assertions in conformance tests
